Fix bugs

+++ notes/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Note, Tag
+import re
 
+regex = re.compile(".*\S.*")
 
 def index(request):
     all_notes = Note.objects.all()
@@ -11,8 +13,8 @@ def create(request):
         note_title = request.POST.get('title')
         note_content = request.POST.get('content')
         tag = request.POST.get('tag')
-        if (note_title != '') and (note_content != ''):
-            if tag != '':
+        if ((regex.match(note_title)) and (len(note_title) <= 20)) and (regex.match(note_content)) and (len(tag) <= 20):
+            if regex.match(tag):
                 Tag.objects.get_or_create(tag=tag)
                 Note.objects.create(title=note_title, content=note_content, tag=Tag.objects.get(tag=tag))
             else:
@@ -29,17 +31,14 @@ def update(request):
         note = Note.objects.get(id=note_id)
         note.title = note_title
         note.content = note_content
-        if note_tag != '':
+        if regex.match(note_tag):
             Tag.objects.get_or_create(tag=note_tag)
-            tag = Tag.objects.get(tag=note_tag)
-            note.tag = tag
+            note.tag = Tag.objects.get(tag=note_tag)
         else:
             note.tag = None
-        if tag_id != '' and len(Note.objects.filter(tag=Tag.objects.get(id=tag_id))) == 1:
+        if (tag_id != '') and (len(Note.objects.filter(tag=Tag.objects.get(id=tag_id))) == 1) and (Tag.objects.get(id=tag_id).tag != note_tag):
             Tag.objects.get(id=tag_id).delete()
-        
         note.save()
-
         return redirect('index')
 
 def delete(request):
@@ -73,15 +72,17 @@ def update_tag(request):
         tag_id = request.POST.get('tag-id')
         tag_content = request.POST.get('tag')
         tag = Tag.objects.get(id=tag_id)
-        if len(Tag.objects.filter(tag=tag_content)) > 1:
+        if not regex.match(tag_content):
+            tag.delete()
+        elif len(Tag.objects.filter(tag=tag_content)) > 0 and tag_content != Tag.objects.get(id=tag_id).tag:
             notes = Note.objects.filter(tag=tag)
             for note in notes:
                 note.tag = Tag.objects.get(tag=tag_content)
+                note.save()
             tag.delete()
         else:
             tag.tag = tag_content
             tag.save()
-        note.save()
         return redirect('tags')
